Initial Commit
The five files uploaded each perform different analysis tasks:
- quench_curve_analysis.py is used to fit fractional population vs
waveguide power to a simulated curve, and extract from this fit the
attenuator voltage at which a pi pulse is obtained.
- waveguidecalibration.py performs a similar function to the file listed
above. This code fractional population vs waveguide power data for many
different frequencies, and calibrates each curve to obtain the proper RF
generator power at each frequency.
- analyzetracesv1.py is the code I use to read in the binary files
output from the digitizer and translate them into something readable. I
use inner products with sine and cosine to obtain the amplitude and
phase at the offset frequency and its second harmonic, as well as the DC
offset for the trace.
- fosoffunctionsv2.py contains a handful of helpful functions for
working with phases and my own personal functions for weighted linear
least squares fitting, as well as weighted average and standard
deviation.
- fosofdatasetanalysis.py will look at the files output by
analyzetracesv1.py and use the functions in fosoffunctionsv2.py to
condense the data into the numbers we really care about. The phases are
averaged in multiple ways, and the output is the fit data for the
lineshape, as well as the averaged phases and their uncertainties. In
addition, any less-important time series data (reference voltages, etc.)
is averaged into chunks that are each 1/10 of the dataset, and output to
their own files. Sometimes, this data is also separated by frequency and
subconfiguration.