Uniformity of Summary and Output Files
Started to make the summary and output files for the
fosofdatasetanalysis.DataSet objects consistent with that of the
FOSOFDataSet objects. This way, I'll be able to use the FOSOFSummary
class on an object of either kind. Still to do: update the 'already
analyzed' method, properly save the output files, and properly read
in/use the new output files with the FOSOFDataSet class.